Added oidc-admin view user-roles
