+++ libs/oidc/admin/data-access/src/lib/modules/user-role/user-role.service.ts
@@ -2,6 +2,7 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 
 import { EnvironmentService } from '@tamu-gisc/common/ngx/environment';
+import { ISimplifiedUserRoleResponse } from '@tamu-gisc/oidc/common';
 
 @Injectable({
   providedIn: 'root'
@@ -10,13 +11,14 @@ export class UserRoleService {
   private resource: string;
 
   constructor(private env: EnvironmentService, private http: HttpClient) {
-    this.resource = this.env.value('api_url') + '/user';
+    this.resource = this.env.value('api_url') + '/user-role';
   }
 
   public getAllUserRoles() {
-    return this.http.get<Array<Partial<any>>>(`${this.resource}/user-role`);
+    return this.http.get<Array<Partial<ISimplifiedUserRoleResponse>>>(`${this.resource}`);
   }
 
+  // TODO: Move this function on the backend from user/role/api to POST user-role
   public insertUserRole(entity) {
     return this.http.post(`${this.resource}/role/api`, entity);
   }
