+++ libs/oidc/admin/ngx/src/lib/modules/authorization/pages/view/view.component.html
@@ -1,1 +1,9 @@
-<p>view works!</p>
+<table *ngFor="let clientGroup of ($userRoles | async)">
+    <th>{{clientGroup[0].client.clientName}}</th>
+    <tr *ngFor="let userRole of clientGroup">
+      <td>{{userRole.role.name}}</td>
+      <td>{{userRole.user.name}}</td>
+      <td>{{userRole.user.email}}</td>
+    </tr>
+  </table>
+
