+++ libs/oidc/common/src/lib/modules/user-role/user-role.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { UserRoleController } from '../../controllers/user-role/user-role.controller';
-import { NewUserRole, NewUserRoleRepo } from '../../oidc-common';
+import { NewUserRoleRepo } from '../../oidc-common';
 import { UserRoleService } from '../../services/user-role/user-role.service';
 
 @Module({
